took out functions login, signin, currentUser and logout from auth.js to directory controllers
