+++ controllers/users/currentUser.js
@@ -1,0 +1,8 @@
+const currentUser = async(req, res, next) => {
+    res.json({
+        email: req.user.email,
+        subscription: req.user.subscription,
+    })
+};
+
+module.exports = currentUser;
