+++ controllers/users/index.js
@@ -1,0 +1,11 @@
+const signup = require("./signup");
+const login = require("./login");
+const currentUser = require("./currentUser");
+const logout = require("./logout");
+
+module.exports = {
+    signup,
+    login,
+    currentUser,
+    logout,
+}
