+++ controllers/users/logout.js
@@ -1,0 +1,9 @@
+const {User} = require("../../models/users");
+
+const logout = async(req, res, next) => {
+    const {_id} = req.user;
+    await User.findByIdAndUpdate(_id, {token: null});
+    res.status(204).send();
+};
+
+module.exports = logout;
